+++ src/layout/footer.tsx
@@ -6,63 +6,107 @@ const Footer = () => {
       display="flex"
       color="#FFFFFF"
       fontSize="16px"
-      lineHeight="23px"
+      fontFamily="Jost"
       fontStyle="normal"
       bgColor=" #353949"
       fontWeight="400px"
       flexDirection="row"
-      // fontFamily="Jost"
       justifyContent="space-between"
       padding="80px 120px 60px 90px"
     >
       <Box width="360px" gap="12px">
         <Image src="/images/hernalytics-full-logo.svg" alt="hernalytics-logo" />
+
         <Stack paddingTop="12px">
-          <Text _hover={{ borderBottom: "1px solid blue" }}>
+          <Text height="24px" _hover={{ borderBottom: "1px solid blue" }}>
             Isale Eko Avenue, Dolphin Estate, Ikoyi, Lagos
           </Text>
-          <Text width="56px" _hover={{ borderBottom: "1px solid blue" }}>
+          <Text
+            width="56px"
+            height="24px"
+            _hover={{ borderBottom: "1px solid blue" }}
+          >
             Nigeria
           </Text>
-          <Text width="140px" _hover={{ borderBottom: "1px solid blue" }}>
+          <Text
+            height="24px"
+            width="140px"
+            _hover={{ borderBottom: "1px solid blue" }}
+          >
             © 2022 Hernalytics
           </Text>
         </Stack>
       </Box>
 
-      <VStack alignItems="left" gap="24px">
-        <Text fontWeight="bold">About</Text>
-        <Text>Our Story</Text>
-        <Text>Blog</Text>
-        <Text>About Hernalytics</Text>
-        <Text width="50px">Videos</Text>
-      </VStack>
-      <VStack alignItems="left" gap="24px">
-        <Text _hover={{ borderBottom: "1px solid blue" }} fontWeight="bold">
-          Support
-        </Text>
-        <Text _hover={{ borderBottom: "1px solid blue" }}>FAQs</Text>
-        <Text _hover={{ borderBottom: "1px solid blue" }}>Privacy Policy</Text>
-        <Text _hover={{ borderBottom: "1px solid blue" }}>
-          Terms of Service
-        </Text>
-      </VStack>
-      <VStack alignItems="left" gap="24px">
-        <Text _hover={{ borderBottom: "1px solid blue" }} fontWeight="bold">
-          Let's chat!
-        </Text>
-        <Text _hover={{ borderBottom: "1px solid blue" }}>
-          hernalytics@gmail.com
-        </Text>
-        <Text _hover={{ borderBottom: "1px solid blue" }}>
-          +234 801 234 5678
-        </Text>
-        <HStack gap="30px">
-          <Image src="/images/instagram.svg" alt="instagram" />
-          <Image src="/images/twitter.svg" alt="twitter" />
-          <Image src="/images/whatsapp.svg" alt="whatsapp" />
-        </HStack>
-      </VStack>
+      <Stack display="flex" flexDirection="row" gap="250px">
+        <VStack alignItems="left" gap="24px">
+          <Text
+            fontWeight="bold"
+            _peerHover={{ borderBottom: "1px solid blue" }}
+          >
+            About
+          </Text>
+          <Text>Our Story</Text>
+          <Text>Blog</Text>
+          <Text>About Hernalytics</Text>
+          <Text width="50px">Videos</Text>
+        </VStack>
+
+        <Stack
+          gap="24px"
+          display="flex"
+          alignItems="left"
+          flexDirection="column"
+        >
+          <Text
+            width="80px"
+            fontWeight="bold"
+            _hover={{ borderBottom: "1px solid blue" }}
+          >
+            Support
+          </Text>
+
+          <Text _hover={{ borderBottom: "1px solid blue" }}>FAQs </Text>
+          <Text _hover={{ borderBottom: "1px solid blue" }} width="140px">
+            Privacy Policy
+          </Text>
+          <Text _hover={{ borderBottom: "1px solid blue" }} width="140px">
+            Terms of Service
+          </Text>
+        </Stack>
+
+        <Stack
+          gap="24px"
+          display="flex"
+          fontSize="16px"
+          alignItems="left"
+          flexDirection="column"
+        >
+          <Text
+            width="80px"
+            height="24px"
+            fontWeight="bold"
+            _hover={{ borderBottom: "1px solid blue" }}
+          >
+            Let's chat!
+          </Text>
+          <Text height="24px" _hover={{ borderBottom: "1px solid blue" }}>
+            hernalytics@gmail.com
+          </Text>
+          <Text
+            width="160px"
+            height="24px"
+            _hover={{ borderBottom: "1px solid blue" }}
+          >
+            +234 801 234 5678
+          </Text>
+          <HStack gap="30px">
+            <Image src="/images/instagram.svg" alt="instagram" />
+            <Image src="/images/twitter.svg" alt="twitter" />
+            <Image src="/images/whatsapp.svg" alt="whatsapp" />
+          </HStack>
+        </Stack>
+      </Stack>
     </Box>
   );
 };
